Lock and unlock actions (#3)
Add lock and unlock action to source view and file content view. If a file is locked it can only be edited by the locking user. But every user is free to remove this lock and edit it again. Also show a warn modal if a user tries to upload and overwrite a locked file.

Co-authored-by: Sebastian Sdorra <sebastian.sdorra@cloudogu.com>
Co-authored-by: René Pfeuffer <rene.pfeuffer@cloudogu.com>